Basic support for 32 drives instead of only 16.

+++ bdos/fsmain.c
@@ -2,13 +2,11 @@
  * fsmain.c - GEMDOS file system notes and misc routines          
  *
  * Copyright (c) 2001 Lineo, Inc.
- *               2002-2008 The EmuTOS development team
+ *               2002 - 2010 The EmuTOS development team
  *
  * This file is distributed under the GPL, version 2 or at your
  * option any later version.  See doc/license.txt for details.
  */
-
-
 
 #include "portab.h"
 #include "asm.h"
@@ -430,7 +428,7 @@ void    xsetdta(char *addr)     /* set transfer address to addr */
 
 long    xsetdrv(int drv)
 {
-    int drvmap ;
+    long drvmap ;
 
     drvmap = Drvmap();
